:art: secretkey 네이밍 수정

+++ src/main/java/animal/meeting/domain/meeting/entity/ProgressingGroup.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ProgressingGroup {
 	private String groupId;
-	private int weightValue;
+	private double weightValue;
 }
